Use root to remove sync file for now
The sudo rules for the admin user running as git don't currently
allow running rm. This results in sudo trying to prompt for a password
and failing due to no tty.

The admin user may run rm as root, though, so we'll use that for
now. We should probably allow rm as the git user explicitly.

/cc @sbryant