Axe 'See Also' section and revise 'Support' section a bit
The links in the See Also section are now up at the top in the
Features section so duplicating here doesn't feel like it's adding
much.

The changes to the Support section are to encourage general bug
reports and feature requests to go to this repository and issues
related to specific GitHub Enterprise setups to go to GitHub
Support.