ghe-backup fails fast if other backup run already in progress
We use ln -sn to synchronize on an in-progress file and avoid races
here.

The command exits non-zero when another in progress backup is
detected. This will generate emails from cron which is useful
information to be alerted to. If backups are running over each
other, the frequency may need to be increased.