Reorder ghe-backup steps for better consistency between data stores
Adjusts the order of backup steps in ghe-backup to reduce potential
data consistency problems between data stores. The new order is:

 - Instance settings
 - License file
 - SSH authorized keys
 - SSH host keys
 - MySQL
 - Redis
 - Git repositories
 - GitHub Pages
 - Elasticsearch

The main change here is moving the MySQL backup step before the
Git repository step. This should prevent cases where repositories
created on the instance *after* Git repositories have been sync'd
but before mysql has been sync'd would result in a missing
repository on the site after restore. With the new order, it's
possible that repositories deleted on the instance *after* the mysql
step but before the Git repositories step will result in a missing
repository but this is a much less severe issue.