Only restore settings when -c given explicitly
The previous behavior would detect if the appliance had ever had a
manual config run and automatically restore settings if not. This
has caused a number of problems in manual testing, some due to bugs
importing settings from an 11.10.34x appliance into a v2.0 appliance
and others due to settings like hostname and SSL being imported into
appliances that are being set up as secondary / staging servers.
It's also just been surprising that settings aren't restored on
subsequent runs.

Making the settings restore an option that must be given explicitly
removes much of the confusion around the auto detection at the
expense of a slightly smoother experience in the perfect case.