how it differs from HA instance
A customer asked why they needed to use the backup-utils utility while we were onsite last week. Their assumption was that because all data is replicated to the High Availability instance the duplication of this data in the form of the backup-utils utility was unnecessary. I wanted to define the separate use cases here so we could point customers to it.