Move in progress detection to separate file with pid
This changes the logic of detecting if a backup is in progress by
writing out the snapshot and PID of the process making the backup. This
improves the situation where a previous backup was killed in a non-clean
way. Clean failures in the sense of a proper exit (also non 0 exit
statuses) or using ctrl-c was already handled by traps.

This just removes the in-progress file if the process is no longer
running after the backup. It then cleans up the in-progress file and
lets the rest of the system clean up the failed backup.

Fixes #99