Specify backup extension
The sed command in OS X requires an extension be passed when using
the in-place flag. While it is possible to specify no extension,
the way this is accomplished differs between the implementations.
It is therefore simpler to specify a valid extension, and remove
the resulting backup file immediately afterwards.