Backup and restore the appliance UUID
When backing up and restoring GitHub Enterprise 2.8.X, we need to copy the UUID over,
which is located in /data/user/common/uuid.

Without this UUID, the appliance will have a different UUID
after the restore, breaking all the existing routes for repositories and leaving
repositories inaccessible after the restore.

No data loss happens. The issue can be easily fixed by manually re-creating the
UUID file after a restore (before re-configuring the appliance), adding the right UUID to it.

This fix is required for customers running GitHub Enterprise 2.8.X (clusters not affected).

It's highly recommended to backup a 2.8.X appliance with backup-utils 2.8.2, that will include
a fix for this.

Customers backing up and restoring older GitHub Enterprise versions are not affected.