Check backup-utils are not being run on GitHub Enterprise host
Customers occasionally configure backup-utils to run on the GitHub Enterprise
host or HA replica. Apart from losing their backup if the appliance is lost,
it leads to running out of disk space issues.

This PR adds a check for this case and returns an informative error to guide the user.