Remove check for git
This introduced a bug, as the && isn't interpreted as part of the ||, so GHE_DISABLE_SSH_MUX was always being set to true.

As git is listed as part of our requirements, and is being used in more and more places in the backup utilities, it should no-longer be optional.