Use sed instead of Ruby
We can't rely on the customer's admin box (which isn't the appliance)
having any actual programming languages on it.  But we can use sed, and
sed is good enough to simulate dirname with a single fork+exec.

This commit also fixes future versions: 2.20.x, 3.x, etc.  All future
versions will be >= 2.19.3, so we just combine the 2.19 check with the
future-versions check.