Run `uniq | sort | uniq`
The data we're sorting has clusters of duplicates in the input, because
`dirname` reduces all repos in the same network (i.e., forks) to the same
network path.  Running `uniq` before `sort` eliminates those duplicates,
which means `sort` requires less CPU and RAM to do its thing.

We still need `uniq` on the output end, because there's no guarantee that
all duplicates in the input are clustered.

I've run tests, and the cost of `uniq` is small enough that it does no
harm if the input has no duplicates at all.