Fix restoring the password pepper for already configured instances
We normally skip settings restore when restoring into an already
configured instance. This is to not overwrite / reset settings
unexpectedly. This is fine for all settings, except for the password
pepper.

The password pepper is associated with the MySQL data and GitHub
passwords used there, so it needs to be restored always together with
the MySQL restore.

This moves the pepper restore to always be done together with the MySQL
restore. We always here update the variable used here since the
`restore-secret` function expects $GHE_RESTORE_SNAPSHOT_PATH to be set.
We had a differently named variable in the MySQL restore with the same
value, so that variable was renamed to match the `restore-secret`
expectation so it can find the backed up password pepper.