Fix "Architecture: all", instead of "any"
"Architecture:" field shows it depends which architecture, if you specify
it as "any" it would be converted on build host arch. "all" means architecture
independs.

Well, this package does not contain any architecture depend files. No .so files,
and /usr/bin/* files are shellscripts.

> $ file ghe-*
> ghe-backup:     Bourne-Again shell script, ASCII text executable
> ghe-host-check: Bourne-Again shell script, ASCII text executable
> ghe-restore:    Bourne-Again shell script, ASCII text executable

So it should be "Architecture: all", not "any".