Use << Here Document redirection instead of <<-
I originally wanted to use the <<- redirection format here to strip
the leading whitespace, but apparently only tab characters are
stripped from the beginning of the lines in a here document; not
spaces. For simplicity sake (and to avoid mixing indentation in our
scripts) I'll just use the regular << redirection syntax to avoid
issues.