Record total runtime of ghe-backup and ghe-restore
This adds the total runtime of ghe-backup and ghe-restore to the
benchmark output.

When investigating benchmark performance issues with customers and
inspecting the benchmark output, it is helpful to know the total runtime
of ghe-backup and ghe-restore as a reference for checking the durations
of individual execution steps. While the stdout output currently shows
the total runtime, customers generally do not store this output on disk,
which makes it difficult to investigate backup-related performance
issues after the fact.

The runtime is recorded regardless of whether or not the commands
completed successfully, as it may still be useful to see how long they
have run when terminated early because of an error or human
intervention.