Remove duplicate definitions in ghe-backup-config (#480)
While investigating an unrelated issue, I noticed that some definitions
in ghe-backup-config are present twice. The reason for this appears to
be a faulty merge conflict resolution [1], as both parent commits [2, 3]
only have a single copy of these definitions but the merge commit has
duplicates. It seems that an unwieldy conflict came up while merging the
progress-indicator with the master branch, as a result of which both the
old and new locations of these definitions were accidentally kept,
causing the duplication.

This removes one copy of each definition to avoid confusion and
potential future bugs.

[1] 1eaf80941c84c21f148ea68783415a920d32ce15
[2] a8e7eca745a738c4c02cba4ce85f51bfb97bfe6b
[3] 8aaccc3c8b6599d591239d22cb789572df05465a