Avoid duplicating test output
A long time ago, we used to track stdout and stderr separately. When
both were combined, the postprocessing instructions of stdout and stderr
were incorrectly merged, resulting in the output file being provided on
stdin twice [1]. As a process cannot take two input streams, one of them
was closed, causing no apparent issue.

However, this triggered a shellcheck issue that was later addressed
improperly [2], which instructed grep to read the output file from both
stdin (by providing the - shorthand for /dev/stdin) and the output file
location (by providing it as another file name argument).

For this reason, the test output would contain the entire test output
twice, once prefixed with “(standard input)” and once with the path to
the output file.

This fixes the issue by simply removing one of the occurrences of the
output file, giving us a single copy of the test output in the case of a
failure without unwanted prefixes in front of every output line.

[1] 2915ec9aa3a11d4936641bdc9f9f43111817047b
[2] 2bc31a99174a9f6ab3de361d3fca18f4bcf2c254