Make test output collapsible
To make the test logs more pleasant to interact with, this makes the
test output of failed test cases collapsible. This avoids cluttering the
test overview with the typically very long output of the individual test
cases, such that users can get a full overview of all tests by default.
If needed, the output of a specific test can easily be expanded in the
GitHub Actions user interface.

As the GitHub Actions user interface already indents the content of
groups, we no longer need to indent it with sed.